requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==3.4.8
cssbeautifier==1.14.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-allauth==0.53.1
django-ckeditor==6.5.1
django-crispy-forms==1.11.1
django-etc==1.3.1
django-hitcount==1.3.5
django-import-export==2.7.1
django-js-asset==2.0.0
django-modeltranslation==0.17.5
django-rename-app==0.1.3
django-widget-tweaks==1.4.12
EditorConfig==0.12.3
et-xmlfile==1.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.7
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.9
pathspec==0.11.0
Pillow==8.1.2
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
tablib==3.2.0
tomli==2.0.1
tqdm==4.64.1
typing-extensions==3.10.0.0
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
